requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.40.0
Deprecated==1.2.15
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
fonttools==4.55.3
gevent==24.11.1
geventhttpclient==2.3.3
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
locust==2.32.6
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
minio==7.2.11
mlflow==2.19.0
mlflow-skinny==2.19.0
msgpack==1.1.0
numpy==2.2.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
starlette==0.41.3
threadpoolctl==3.5.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.1
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
